pulled functions out into a utils folder so that reducer would be more readeable and refactorable
