+++ client/reducers/mainReducer.js
@@ -1,130 +1,12 @@
 import * as types from '../constants/actionTypes';
 import clone from 'clone';
-function easterEgg() {
-  const audio = new Audio(
-    'https://iringtone.net/rington/file?id=8454&type=sound&name=mp3'
-  );
-  audio.play();
-
-  const app = document.getElementById('app');
-  app.classList.add('secret');
-
-  for (let i = 0; i <= 1000; i += 1) {
-    const particle = document.createElement('i');
-    particle.classList.add('particle');
-    app.appendChild(particle);
-  }
-
-  const colorGen = () => {
-    const r = Math.floor(Math.random() * 256);
-    const g = Math.floor(Math.random() * 256);
-    const b = Math.floor(Math.random() * 256);
-
-    return 'rgb(' + r + ',' + g + ',' + b + ')';
-  };
-
-  setInterval(() => {
-    document.querySelector(
-      '.navbar'
-    ).style.backgroundColor = `${colorGen()}`;
-
-    const buttons = document.getElementsByTagName('button');
-    for (let i = 0; i < buttons.length; i += 1) {
-      buttons[i].style.backgroundColor = `${colorGen()}`;
-    }
-
-    const navIcons = document.getElementsByClassName('fas');
-    for (let i = 0; i < navIcons.length; i += 1) {
-      navIcons[i].style.color = `${colorGen()}`;
-    }
-
-    const icons = document.getElementsByClassName('far');
-    for (let i = 0; i < icons.length; i += 1) {
-      icons[i].style.color = `${colorGen()}`;
-    }
-
-    const nodeBase = document.getElementsByClassName('nodeBase');
-    for (let i = 0; i < nodeBase.length; i += 1) {
-      nodeBase[i].style.fill = `${colorGen()}`;
-    }
-
-    const leafNodeBase = document.getElementsByClassName('leafNodeBase');
-    for (let i = 0; i < leafNodeBase.length; i += 1) {
-      leafNodeBase[i].style.fill = `${colorGen()}`;
-    }
-  }, 100);
-}
-
-function DoublyLinkedList(value) {
-  this.value = value;
-  this.prev = null;
-  this.next = null;
-}
-
-const getCircularReplacer = () => {
-  const seen = new WeakSet();
-  return (key, value) => {
-    if (typeof value === 'object' && value !== null) {
-      if (seen.has(value)) {
-        return;
-      }
-      seen.add(value);
-    }
-    return value;
-  };
-};
-function deleteChildrenInNameAndCodeLinkedToComponentId(node, obj) {
-  if (node instanceof Object) {
-    delete obj[node]
-    return obj;
-  }
-  node.forEach(childNode => {
-    delete obj[childNode];
-    if (childNode.children) {
-      delete obj[childNode];
-      deleteChildrenInNameAndCodeLinkedToComponentId(childNode.children, obj);
-    }
-  });
-  return obj;
-}
-const findAndDeleteInCurrentComponent = (tree, currentComponent, parent) => {
-  if (tree.componentId === parent.componentId) {
-    for (let i = 0; i < tree.children.length; i++) {
-      if (tree.children[i].componentId === currentComponent.componentId) {
-        tree.children.splice(i, 1);
-        parent.children = tree.children.slice();
-        return;
-      }
-    }
-  } else if (tree.children) {
-    tree.children.forEach(child => {
-      return findAndDeleteInCurrentComponent(child, currentComponent, parent);
-    });
-  }
-};
-function sendToLocalStorage(nameAndCodeLinkedToComponentId, data, currentComponent, history, lastId) {
-  if (nameAndCodeLinkedToComponentId) {
-    localStorage.setObj('nameAndCodeLinkedToComponentId', nameAndCodeLinkedToComponentId);
-  }
-  if (data) {
-    localStorage.setObj('data', data);
-  }
-  if (currentComponent) {
-    localStorage.setObj('currentComponent', currentComponent);
-  }
-  if (history) {
-    localStorage.setObj('history', history);
-  }
-  if (lastId) {
-    localStorage.setObj('lastId', lastId);
-  }
-}
-Storage.prototype.setObj = function (key, obj) {
-  return this.setItem(key, JSON.stringify(obj, getCircularReplacer()));
-};
-Storage.prototype.getObj = function (key) {
-  return JSON.parse(this.getItem(key));
-};
+import {
+  deleteChildrenInNameAndCodeLinkedToComponentId,
+  findAndDeleteInCurrentComponent
+} from './utils/deleteNodeFuncs';
+import { easterEgg } from './utils/easterEgg'
+import { sendToLocalStorage } from './utils/localStorage'
+import { updateTree, DoublyLinkedList } from './utils/updateTree'
 
 const appComponent = {
   name: 'App',
@@ -146,63 +28,6 @@ const initialState = {
   templates: [],
   orientation: 'vertical',
   toggleFileTree: true
-};
-
-const updateTree = (state, currentComponent) => {
-  let defaultNameCount;
-  // check if current component has a name
-  const data = clone(state.data);
-  if (!currentComponent.name) {
-    defaultNameCount = state.defaultNameCount + 1;
-    currentComponent.name = `Component${defaultNameCount}`;
-  }
-  // check if any child has empty name, then change it to 'DEFAUL NAME'
-  let children = currentComponent.children
-  if (children) {
-    for (let child of children) {
-      if (child.name === '') {
-        defaultNameCount = defaultNameCount
-          ? defaultNameCount + 1
-          : state.defaultNameCount + 1
-      }
-    }
-  } else {
-    children = currentComponent
-    children.name = currentComponent.name;
-  }
-  (function findComponentAndUpdate(tree, currentComponent) {
-    if (tree.componentId === currentComponent.componentId) {
-      tree.name = currentComponent.name;
-      tree.isContainer = currentComponent.isContainer;
-      tree.children = currentComponent.children
-      return;
-    }
-    if (tree.children) {
-      tree.children.forEach(child => {
-        return findComponentAndUpdate(child, currentComponent);
-      });
-    }
-  }(data, currentComponent));
-  const preHistory = clone(state.history);
-  const history = new DoublyLinkedList(
-    {
-      data,
-      currentComponent,
-      nameAndCodeLinkedToComponentId: state.nameAndCodeLinkedToComponentId,
-      lastId: state.lastId,
-      defaultNameCount: state.defaultNameCount
-    }
-  );
-  preHistory.next = history;
-  history.prev = preHistory;
-  return {
-    data,
-    currentComponent,
-    history,
-    defaultNameCount: defaultNameCount
-      ? defaultNameCount
-      : state.defaultNameCount
-  };
 };
 
 const mainReducer = (state = initialState, action) => {
